+++ Unstable/StarFighter/StarFighter/Evolution.h
@@ -26,9 +26,9 @@ struct Evolution
 	int m_mutation;
 	int m_crossover;
 	int m_gen;
-	int m_dna_input_a[DNA_LENGTH];
-	int m_dna_input_b[DNA_LENGTH];
-	int m_dna_output[DNA_LENGTH];
+	uint8_t m_dna_input_a[DNA_LENGTH];
+	uint8_t m_dna_input_b[DNA_LENGTH];
+	uint8_t m_dna_output[DNA_LENGTH];
 };
 
 #endif // EVOLUTION_H_INCLUDED
